+++ Tests/Functional/BackEnd/AbstractEventMailFormTest.php
@@ -38,7 +38,8 @@ final class AbstractEventMailFormTest extends FunctionalTestCase
     {
         parent::setUp();
 
-        $this->setUpBackendUserFromFixture(1);
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/BackEndUser.csv');
+        $this->setUpBackendUser(1);
         $this->initializeBackEndLanguage();
 
         $this->email = $this->createEmailMock();